backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.0
redis==5.0.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Data validation and settings
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP client
httpx==0.25.2
aiofiles==23.2.1

# Google OAuth
google-auth==2.23.4
google-auth-oauthlib==1.1.0

# Crypto trading
ccxt==4.1.77
python-binance==1.0.19

# Data processing
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2

# Technical analysis
ta==0.10.2

# Utilities
qrcode==7.4.2
Pillow==10.1.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
